+++ ProductService/src/main/java/com/example/ProductService/dtos/CreateProductRequestDto.java
@@ -7,5 +7,5 @@ public class CreateProductRequestDto {
     String name;
     String description;
     String category;
-    int price;
+    Double price;
 }
